requirements.txt
# Web Framework
fastapi>=0.109.0
uvicorn>=0.27.0

# UI
streamlit>=1.31.0

# LLM
openai>=1.12.0

# HTTP Client
httpx>=0.26.0

# Data Validation
pydantic>=2.6.0
pydantic-settings>=2.1.0

# Environment Variables
python-dotenv>=1.0.1

# Retry Logic
tenacity>=8.2.3

# Async Support
anyio>=4.0.0
asyncio-throttle>=1.0.2
